Brad's attempt to cut to 5 pages
So here is most of my attempt to cut this paper to 5 pages.
I amde it to about 5.7 pages.  I could go after it some more,
but this is a good first draft point to check and see if we need
to add back anything I took out.

I also added the IEEEabbrv.bib file since I couldn't build
without it.